requirements.txt
fastapi
uvicorn[standard]
Pillow
opencv-python
torch
numpy
deepface
ultralytics
super-image
python-multipart
